requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
